Do setup/tear-down for all tests
